+++ packages/health_connector_hc_android/test/unit_tests/src/health_connector_hc_client_test.dart
@@ -870,6 +870,25 @@ void main() {
               );
             },
           );
+
+          group(
+            'readCharacteristic',
+            () {
+              test(
+                'throws UnsupportedOperationException',
+                () {
+                  expect(
+                    () => client.readCharacteristic(
+                      HealthCharacteristicType.biologicalSex,
+                    ),
+                    throwsA(
+                      isA<UnsupportedOperationException>(),
+                    ),
+                  );
+                },
+              );
+            },
+          );
         },
       );
     },