+++ packages/health_connector_hc_android/test/unit_tests/src/mappers/permission_mappers/permission_mapper_test.dart
@@ -64,6 +64,23 @@ void main() {
               );
             },
           );
+
+          test(
+            'throws UnsupportedOperationException for '
+            'HealthCharacteristicPermission',
+            () {
+              const permission = HealthCharacteristicPermission(
+                HealthCharacteristicType.biologicalSex,
+              );
+
+              expect(
+                () => permission.toDto(),
+                throwsA(
+                  isA<UnsupportedOperationException>(),
+                ),
+              );
+            },
+          );
         },
       );
     },